+++ vanilla/src/crowmap1_21_1/java/agency/highlysuspect/crowmap/craftful/mixin/MapItemMixin.java
@@ -1,0 +1,20 @@
+package agency.highlysuspect.crowmap.craftful.mixin;
+
+import net.minecraft.world.item.MapItem;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
+
+@Mixin(MapItem.class)
+public class MapItemMixin {
+	@ModifyVariable(
+		at = @At("HEAD"),
+		method = "inventoryTick",
+		argsOnly = true
+	)
+	public boolean onEntityTickPre(boolean isHeld) {
+		//this bool controls whether the map is held in the player's hand or not
+		//so... to make the map always update, time to always return true :P
+		return true;
+	}
+}